add implementation to OutputCreator
OutputCreator can now be basically used. Still misses the addition of
the nouns, and complements, which is pretty important!
sentences.json - structure changed, now using JSONArrays to store
sentences (easier to access randomly)
Generation of randomNumber to access a sentence is now done in
OutputCreator, and not in the Enum anymore. Reason: more dynamic, and
saves hand-work.