Time management in CanteenData + bugs + ..
constructor now needs an enum from CanteenNames,
time managment is supported with constructor -> 0 for today, 1 for
tomorrow, -1 for yesterday. I will add isOpen attribute to the lines and
mensa data so it can easily be checked if the lines/mensa is open. A
mensa is open when at least one line is open.
This needs some fine tuning, so i'm not ready yet! :D